requirements.txt
fastapi
uvicorn
python-dotenv
gspread
oauth2client
pyotp
qrcode
